fixed business model

+++ Business/Model/CustomerDetails.cs
@@ -9,7 +9,7 @@ namespace Business.Model
             RecentTransactions = new List<TransactionDetails>();
         }
 
-        public decimal CustomerId { get; set; }
+        public int CustomerId { get; set; }
         public string Email { get; set; }
         public string Name { get; set; }
         public decimal Phone { get; set; }
